refactor(sync): merge bead model into beads store (bead.ts + beads-store.ts -> beads.ts)
Combine the beads domain model (Bead types, parseBeads, blockerBeadIds,
parentBeadId) with the bd CLI store (BeadsStore/ReverseBeadsStore, BdCliStore)
into a single beads.ts — the beads-side counterpart to github.ts. Update all
imports. Behavior-preserving; 49/49 script tests pass.
